refactor: reexport internal components without private properties
