+++ docs/components/ColorToken.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react"
 
-import { Button } from "@/components/Actions/Button"
 import { cn } from "@/lib/utils"
 import { CopyIcon } from "lucide-react"
+import { Button } from "../../lib/components/Actions/Button/internal"
 
 type Props = {
   name: string
