+++ lib/components/Actions/exports.tsx
@@ -1,13 +1,7 @@
 import { Component } from "@/lib/component"
-import { removeInternalProps } from "@/lib/internal-props.tsx"
 import { Link as LinkComponent } from "./Link"
 
-// Button
-import { Button as ButtonComponent } from "./Button"
-
-export const Button = removeInternalProps(ButtonComponent)
-export * from "./Button"
-//
+export * from "./Button/"
 
 export type { LinkProps } from "./Link"
 export const Link = Component(
