+++ lib/experimental/Navigation/Dropdown/index.tsx
@@ -1,114 +1,37 @@
-import { Button, ButtonProps } from "@/components/Actions/Button"
-import { Icon, IconType } from "@/components/Utilities/Icon"
-import { AvatarVariant } from "@/experimental/Information/Avatars/Avatar"
+import { Button } from "@/components/Actions/Button"
+import { Icon } from "@/components/Utilities/Icon"
 import { EllipsisHorizontal } from "@/icons/app"
-import { Link } from "@/lib/linkHandler"
-import { cn } from "@/lib/utils"
+import { Link } from "@/lib/linkHandler.tsx"
+import { cn } from "@/lib/utils.ts"
 import {
   Drawer,
   DrawerContent,
   DrawerOverlay,
   DrawerTrigger,
-} from "@/ui/drawer"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/ui/dropdown-menu"
+} from "@/ui/drawer.tsx"
 import { useState } from "react"
-import { NavigationItem } from "../utils"
-import { DropdownItemContent } from "./DropdownItem"
+import { DropdownItemContent } from "./DropdownItem.tsx"
+import {
+  DropdownInternal,
+  DropdownInternalProps,
+  DropdownItem,
+  DropdownItemObject,
+} from "./internal.tsx"
 
-export type DropdownItem = DropdownItemObject | "separator"
+const privateProps = ["align"] as const
 
-export type DropdownItemObject = NavigationItem & {
-  onClick?: () => void
-  icon?: IconType
-  description?: string
-  critical?: boolean
-  avatar?: AvatarVariant
+type DropdownProps = Omit<DropdownInternalProps, (typeof privateProps)[number]>
+
+export const Dropdown = (props: DropdownProps) => {
+  const publicProps = privateProps.reduce((acc, key) => {
+    const { [key]: _, ...rest } = acc
+    return rest
+  }, props as DropdownInternalProps)
+
+  return <DropdownInternal {...publicProps} />
 }
 
-type DropdownProps = {
-  items: DropdownItem[]
-  icon?: IconType
-  size?: ButtonProps["size"]
-  children?: React.ReactNode
-  internalAlign?: "start" | "end"
-}
-
-const DropdownItem = ({ item }: { item: DropdownItemObject }) => {
-  const { label: _label, ...props } = item
-
-  const itemClass = cn(
-    "flex items-start gap-1.5 w-full",
-    item.critical && "text-f1-foreground-critical"
-  )
-
-  return (
-    <DropdownMenuItem
-      asChild
-      onClick={(e) => {
-        e.stopPropagation()
-        item.onClick?.()
-      }}
-      className={itemClass}
-    >
-      {item.href ? (
-        <Link
-          href={item.href}
-          className={cn(
-            itemClass,
-            "text-f1-foreground no-underline hover:cursor-pointer"
-          )}
-          {...props}
-        >
-          <DropdownItemContent item={item} />
-        </Link>
-      ) : (
-        <div className={itemClass}>
-          <DropdownItemContent item={item} />
-        </div>
-      )}
-    </DropdownMenuItem>
-  )
-}
-
-export function Dropdown({
-  items,
-  icon = EllipsisHorizontal,
-  internalAlign = "start",
-  size,
-  children,
-}: DropdownProps) {
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        {children || (
-          <Button
-            hideLabel
-            icon={icon}
-            size={size}
-            label="..."
-            round
-            variant="outline"
-          />
-        )}
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align={internalAlign}>
-        {items.map((item, index) =>
-          item === "separator" ? (
-            <DropdownMenuSeparator key={index} />
-          ) : (
-            <DropdownItem key={index} item={item} />
-          )
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
+export type { DropdownItem, DropdownItemObject }
 
 export const MobileDropdown = ({ items, children }: DropdownProps) => {
   const [open, setOpen] = useState(false)
