+++ lib/experimental/Utilities/PrivateBox/index.stories.tsx
@@ -1,5 +1,5 @@
 // Replace your-framework with the name of your framework
-import { Button } from "@/../lib/components/Actions/Button"
+import { Button } from "@/components/Actions/Button"
 import { usePrivacyMode } from "@/lib/privacyMode"
 import type { Meta, StoryObj } from "@storybook/react"
 import { PrivateBox } from "."
